修改了所有的接口请求路径，添加试卷问题的处理接口

+++ src/main/java/xyz/bibiyes/goodlearnai/entity/Question.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @TableName("question")
 public class Question {
     @TableId(type = IdType.AUTO)
-    private Integer questionId;
+    private Long questionId;
     private String questionTitle;
     private String questionContent;
     private LocalDateTime creationDate;
